src/base64.c (bin_to_base64): Fix an infinite loop.
* src/base64.c (bin_to_base64): Change type of the second argument
  to int to prevent an infinite loop due to wrapping of the size_t
  value.
* src/base64.h: Update.
* src/key-func.c (guile_ssh_public_key_to_string): Change type of
  `key_len'.
* NEWS: Update.

+++ src/base64.c
@@ -96,7 +96,7 @@ _bin_to_base64(unsigned char *dest,
 
    Return the converted string. */
 unsigned char *
-bin_to_base64(const unsigned char *source, size_t len)
+bin_to_base64(const unsigned char *source, int len)
 {
   unsigned char *base64;
   unsigned char *ptr;
